PROD-130 프로필 전환 캐시 갱신 경로 정리
