--- apps/web/src/lib/profileStateChanged.ts
@@ -1,1 +1,0 @@
-export type ProfileStateChangedReason = 'profile-selected' | 'profile-created';
